+++ packages/quantic/force-app/main/default/lwc/quanticSort/quanticSort.js
@@ -1,6 +1,11 @@
 import {LightningElement, track, api} from 'lwc';
 import {registerComponentForInit, initializeWithHeadless} from 'c/quanticHeadlessLoader';
 
+import sortBy from '@salesforce/label/c.quantic_SortBy';
+import relevancy from '@salesforce/label/c.quantic_Relevancy';
+import newest from '@salesforce/label/c.quantic_Newest';
+import oldest from '@salesforce/label/c.quantic_Oldest';
+
 export default class QuanticSort extends LightningElement {
   @track state = {};
 
@@ -11,6 +16,13 @@ export default class QuanticSort extends LightningElement {
   sort;
   /** @type {import("coveo").Unsubscribe} */
   unsubscribe;
+
+  labels = {
+    sortBy,
+    relevancy,
+    newest,
+    oldest
+  }
 
   connectedCallback() {
     registerComponentForInit(this, this.engineId);
@@ -75,9 +87,9 @@ export default class QuanticSort extends LightningElement {
 
   get options() {
     return [
-      {label: 'Relevancy', value: 'relevancy'},
-      {label: 'Newest', value: 'newest'},
-      {label: 'Oldest', value: 'oldest'},
+      {label: this.labels.relevancy, value: 'relevancy'},
+      {label: this.labels.newest, value: 'newest'},
+      {label: this.labels.oldest, value: 'oldest'},
     ];
   }